+++ src/main/java/com/kelog/kelog/controller/CommentController.java
@@ -1,7 +1,7 @@
 package com.kelog.kelog.controller;
 
 import com.kelog.kelog.request.CommentRequestDto;
-import com.kelog.kelog.response.ResponseDto;
+import com.kelog.kelog.controller.response.ResponseDto;
 import com.kelog.kelog.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
