+++ src/main/java/com/kelog/kelog/controller/HeartController.java
@@ -1,7 +1,6 @@
 package com.kelog.kelog.controller;
 
-import com.kelog.kelog.repository.HeartRepository;
-import com.kelog.kelog.response.ResponseDto;
+import com.kelog.kelog.controller.response.ResponseDto;
 import com.kelog.kelog.service.HeartService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
