+++ src/main/java/com/kelog/kelog/controller/MemberController.java
@@ -3,7 +3,7 @@ package com.kelog.kelog.controller;
 import com.kelog.kelog.request.IdCheckDto;
 import com.kelog.kelog.request.LoginDto;
 import com.kelog.kelog.request.SignUpRequestDto;
-import com.kelog.kelog.response.MemberResponseDto;
+import com.kelog.kelog.controller.response.MemberResponseDto;
 import com.kelog.kelog.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
