+++ src/main/java/com/kelog/kelog/controller/response/ResponseDto.java
@@ -1,4 +1,4 @@
-package com.kelog.kelog.response;
+package com.kelog.kelog.controller.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
