+++ src/main/java/com/kelog/kelog/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.kelog.kelog.exception;
 
-import com.kelog.kelog.response.ResponseDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
