+++ src/main/java/com/kelog/kelog/security/jwt/AccessDeniedHandlerException.java
@@ -1,7 +1,7 @@
 package com.kelog.kelog.security.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kelog.kelog.response.ResponseDto;
+import com.kelog.kelog.controller.response.ResponseDto;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
