+++ src/main/java/com/kelog/kelog/service/PostService.java
@@ -6,7 +6,10 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.util.IOUtils;
-import com.kelog.kelog.domain.Heart;
+import com.kelog.kelog.controller.response.MemberResponseDto;
+import com.kelog.kelog.controller.response.PostAllByResponseDto;
+import com.kelog.kelog.controller.response.PostResponseDto;
+import com.kelog.kelog.controller.response.ResponseDto;
 import com.kelog.kelog.domain.Member;
 import com.kelog.kelog.domain.Post;
 import com.kelog.kelog.domain.Tags;
@@ -23,7 +26,6 @@ import com.kelog.kelog.security.jwt.TokenProvider;
 import com.kelog.kelog.shared.CommonUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.aspectj.weaver.Lint;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -37,7 +39,6 @@ import java.time.LocalDate;
 import java.util.List;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Optional;
 
 import static java.util.stream.Collectors.toList;
